EvaluateInputTensor Callback moved to mnist_tfrecord.py
