+++ examples/mnist_tfrecord.py
@@ -42,7 +42,7 @@ import tensorflow as tf
 import keras
 from keras import backend as K
 from keras import layers
-from keras.callbacks import EvaluateInputTensor
+from keras.callbacks import Callback
 
 from tensorflow.contrib.learn.python.learn.datasets import mnist
 
@@ -51,6 +51,53 @@ if K.backend() != 'tensorflow':
                        'TensorFlow backend, '
                        'because it requires TFRecords, which '
                        'are not supported on other platforms.')
+
+
+class EvaluateInputTensor(Callback):
+    """ Validate a model which does not expect external numpy data during training.
+
+    Keras does not expect external numpy data at training time, and thus cannot
+    accept numpy arrays for validation when all of a Keras Model's
+    `Input(input_tensor)` layers are provided an  `input_tensor` parameter,
+    and the call to `Model.compile(target_tensors)` defines all `target_tensors`.
+    Instead, create a second model for validation which is also configured
+    with input tensors and add it to the `EvaluateInputTensor` callback
+    to perform validation.
+
+    It is recommended that this callback be the first in the list of callbacks
+    because it defines the validation variables required by many other callbacks,
+    and Callbacks are made in order.
+
+    # Arguments
+        model: Keras model on which to call model.evaluate().
+        steps: Integer or `None`.
+            Total number of steps (batches of samples)
+            before declaring the evaluation round finished.
+            Ignored with the default value of `None`.
+    """
+
+    def __init__(self, model, steps, metrics_prefix='val', verbose=1):
+        # parameter of callbacks passed during initialization
+        # pass evalation mode directly
+        super(EvaluateInputTensor, self).__init__()
+        self.val_model = model
+        self.num_steps = steps
+        self.verbose = verbose
+        self.metrics_prefix = metrics_prefix
+
+    def on_epoch_end(self, epoch, logs={}):
+        self.val_model.set_weights(self.model.get_weights())
+        results = self.val_model.evaluate(None, None, steps=int(self.num_steps),
+                                          verbose=self.verbose)
+        metrics_str = '\n'
+        for result, name in zip(results, self.val_model.metrics_names):
+            metric_name = self.metrics_prefix + '_' + name
+            logs[metric_name] = result
+            if self.verbose > 0:
+                metrics_str = metrics_str + metric_name + ': ' + str(result) + ' '
+
+        if self.verbose > 0:
+            print(metrics_str)
 
 
 def cnn_layers(x_train_input):
